requirement.txt
duckduckgo_search          8.1.1
google-api-core            2.28.1
google-api-python-client   2.186.0
groq                       0.33.0
langchain                  0.3.27
langchain-community        0.3.31
python-dotenv              1.2.1
langchain-huggingface      1.0.0
langchain-google-community 2.0.10
streamlit                  1.51.0